+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [unreleased]
+## [v1.9] - 2026-01-11
+
+## Added
+- Added a layout editor that allows for editing or adding to your layouts after they are created.
 
 ## Fixed
 - When merging layouts, labels on tiles are now preserved correctly.
+- Fixed an issue where when merging layouts there was a chance that layouts would be removed even if the merge failed.
+- Scroll speed in the GUI should now be more responsive.
 
 ## [v1.8] - 2026-01-11